pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx[socks]>=0.27.0",
    "pydantic>=2.9.0",
    "pywebview>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
llm-council = "desktop:main"
